+++ lib/ig_markets/cli/activities_command.rb
@@ -5,7 +5,7 @@ module IGMarkets
       desc 'activities', 'Prints account activities'
 
       option :days, type: :numeric, required: true, desc: 'The number of days to print account activities for'
-      option :start_date, desc: 'The start date to print account activities from, in the format \'YYYY-MM-DD\''
+      option :start_date, desc: 'The start date to print account activities from, format: yyyy-mm-dd'
 
       def activities
         self.class.begin_session(options) do |dealing_platform|
@@ -23,7 +23,7 @@ module IGMarkets
 
           dealing_platform.account.activities_in_date_range start_date, start_date + days.to_i
         else
-          dealing_platform.account.recent_activities seconds(days)
+          dealing_platform.account.recent_activities days
         end
       end
     end